+++ docs/changes.rst
@@ -37,6 +37,7 @@ Released on ? 2015.
 * Clarified terminology on strings needing action and not translated strings.
 * Support for Python 3.
 * Dropped support for Django 1.7.
+* Dropped dependency on msginit for creating new Gettext po files.
 
 weblate 2.4
 -----------