+++ docs/changes.rst
@@ -39,6 +39,7 @@ Released on ? 2015.
 * Dropped support for Django 1.7.
 * Dropped dependency on msginit for creating new Gettext po files.
 * Added configurable dashboard views.
+* Improved notifications on parse erorrs.
 
 weblate 2.4
 -----------